class作成・オブジェクト定義WIP
