+++ tsconfig.json
@@ -22,6 +22,7 @@
       ]
     },
     "types": [
+      "types",
       "@types/node",
       "@nuxt/types",
       "@nuxt/content",
